_embed refactored to base, removed from nonpar
